hyper_parameter_tuning/requirements.txt
appnope==0.1.0
autopep8==1.3.1
bleach==2.0.0
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.2
future==0.16.0
h5py==2.7.0
html5lib==0.999999999
hyperas==0.4
hyperopt==0.1
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.6
keras-vis==0.4.1
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nose==1.3.7
notebook==5.0.0
numpy==1.13.1
olefile==0.44
p5py==1.0.0
pandas==0.20.1
pandocfilters==1.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.1
pycodestyle==2.3.1
Pygments==2.2.0
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorflow==1.1.0
terminado==0.6
testpath==0.3
Theano==0.9.0
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
xgboost==0.6a2
